requirements.txt
flask
pandas
tweepy
textblob
wordcloud
matplotlib
nltk
python-dotenv
